Process markdown in atom feed, and some improvements.
The change which fixes the problem of markdown in the feed is the markdownify
filter. I don't know why this is necessary here and not in the site wide atom
feed. Also, I expect this will only work with markdown - that is if I were to
use some other source format it might break.